consolidated tests to one file
